pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opteryx-data"
version = "0.0.0"
description = "Opteryx Data Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.101.0",
    "uvicorn[standard]==0.23.2",
    "python-jose[cryptography]==3.3.0",
    "cryptography>=40.0.0",
    "requests>=2.30.0",
    "python-multipart>=0.0.6",
    "google-cloud-secret-manager>=2.19.0",
    "google-cloud-firestore>=2.10.0",
    "google-cloud-tasks>=2.11.0",
    "argon2-cffi>=21.3.0",
    "orjson>=3.9.0",
    "orso",
    "opteryx==0.26.1"
]

[tool.isort]
profile = "black"
extend_skip_glob = ["tests/**"]
skip_gitignore = true
line_length = 100
multi_line_output = 9
force_single_line = true
float_to_top = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
indent-width = 4
target-version = 'py313'

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 100